done excursions count added to employee, some new PAST excursions added to db

+++ src/main/java/com/softserve/academy/museum/service/EmployeeService.java
@@ -27,6 +27,6 @@ public interface EmployeeService {
 
     long getWorkTime(int id, LocalDateTime from, LocalDateTime to);
 
-    int getExcursionCount(int id);
+    long getExcursionCount(int id);
 
 }
